K8S module refactor - separate plumbing code from k8s logic
