+++ h2o-k8s/src/main/java/water/k8s/KubernetesEmbeddedConfig.java
@@ -63,7 +63,7 @@ public class KubernetesEmbeddedConfig extends AbstractEmbeddedH2OConfig {
 
     @Override
     public boolean disableNonLeaderNodeAccess() {
-        return KubernetesEmbeddedConfigProvider.isRunningOnKubernetes();
+        return H2OCluster.isRunningOnKubernetes();
     }
 
 }
